Remove extraneous features
This strips out the concept of using slots in each slab, using an Arena
for multiple Counter objects, and an ArenaManager object. The goal is to
strip the code down to the bare essentials so that the raw performance
of the idea could be evaluated.

With that in mind, the Arena is also... different. It is not a variable
sized object that cannot be created normally. This hack avoids the need
for an extra pointer indirection in an std::vector, but oh does that
code make me uneasy.

This diff also fixes the benchmarks. It turns out `std::shared_ptr` has
optimizations for assigning a `std::shared_ptr` back to itself, so the
benchmark now resets the pointers in the per-thread buffers before
assignment.