Fix a race condition
The issue was that one thread could be decrementing a slab counter to 0
while another thread was recording its usage on that slab counter.

The conceptual solution starts with having the slab counter have an
uninitialized state (any value below 0), and then require that all
increments to the counter fail if they detect the counter is
uninitialized. If the counter is uninitialized, then as long as the
thread is able to register the usage of the CPU with the Arena, it is
allowed to set the counter to an initialized value.

The thread that decrements the counter to 0 will attempt to place the
Slab counter in an uninitialized state, and if it succeeds, it will
always unmark the CPU. After the Slab counter is placed in an
unitialized state, if a thread is attempting to increment the Slab
counter; it will be unable to mark the CPU usage with the Arena until
the decrementing thread has finished.